+++ proteus-lib/src/audio/mod.rs
@@ -1,0 +1,2 @@
+pub mod buffer;
+pub mod samples;
