+++ proteus-lib/src/audio/samples.rs
@@ -1,0 +1,12 @@
+use rodio::{buffer::SamplesBuffer, Source};
+
+pub fn clone_samples_buffer(
+    buffer: SamplesBuffer<f32>,
+) -> (SamplesBuffer<f32>, SamplesBuffer<f32>) {
+    let sample_rate = buffer.sample_rate();
+    let buffered = buffer.buffered();
+    let vector_samples = buffered.clone().into_iter().collect::<Vec<f32>>();
+    let clone1 = SamplesBuffer::new(buffered.channels(), sample_rate, vector_samples.clone());
+    let clone2 = SamplesBuffer::new(buffered.channels(), sample_rate, vector_samples);
+    (clone1, clone2)
+}
