+++ proteus-lib/src/dsp/mod.rs
@@ -1,5 +1,4 @@
 pub mod convolution;
-pub mod effects;
 pub mod impulse_response;
-pub mod spring_impulse_response;
 pub mod reverb;
+pub mod spring_impulse_response;
