+++ proteus-lib/src/dsp/spring_impulse_response.rs
