+++ proteus-lib/src/lib.rs
@@ -1,17 +1,14 @@
 //! # Proteus Audio Library
 //!
 //! This library provides the core audio playback and processing functionalities for the Proteus audio player.
-//! It includes modules for audio playback, effects, metadata handling, and more.
+//! It includes modules for audio playback, DSP effects, metadata handling, and more.
 
-mod buffer;
-pub mod effects;
-pub mod info;
+pub mod audio;
+pub mod container;
+pub mod diagnostics;
+pub mod dsp;
 pub mod peaks;
-pub mod player;
-mod player_engine;
-pub mod prot;
-pub mod reporter;
+pub mod playback;
 pub mod test_data;
-pub mod timer;
-mod tools;
+pub mod tools;
 mod track;
